src/pyproject.toml
[project]
name = "gcp-cloud-run-container"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.1",
    "google-cloud-secret-manager>=2.24.0",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "ruff>=0.12.7",
]
